Add Repository

+++ AdopPix.Models/UserProfile.cs
@@ -15,7 +15,7 @@ namespace AdopPix.Models
         public string coverName { get; set; }
         public string fname { get; set; }
         public string lname { get; set; }
-        public int money { get; set; }
+        public double money { get; set; }
         public DateTime birthDate { get; set; }
         public int point { get; set; }
         public int rank { get; set; }
